+++ trunk/Construcción/UPC.Proyecto.SISPPAFUT/UPC.Proyecto.SISPPAFUT/frmModuloEntrenamiento.cs
@@ -210,7 +210,7 @@ namespace UPC.Proyecto.SISPPAFUT
                                         objPartidoPronosticadoBE.C_Local_Pts = objPartidoPronosticadoBE.C_Local_Pts + 1;
                             }
                         }
-                        objPartidoPronosticadoBE.C_Local_PosLiga = objTablaBC.ConsultarPosicionEquipoTabla(codLiga, codEquipoL);
+                        objPartidoPronosticadoBE.C_Local_PosLiga = (int)objTablaBC.ConsultarPosicionEquipoTabla(codLiga, codEquipoL);
                         objPartidoPronosticadoBE.C_Local_PosRankMund = objRankingBC.obtener_PosRanking(cDto.Fecha.Year, cDto.Fecha.Month, paisL, cDto.Equipo_local);
                         objPartidoPronosticadoBE.C_Local_PromEdad = objEquipoBC.obtener_PromedioEquipoTitular(codEquipoL, codLiga);
                         objPartidoPronosticadoBE.C_Local_QExpulsados = objEquipoBC.obtener_CantidadExpulsadosUltimoPartido(codUltimoPartidoLocal, codEquipoL, codLiga);
@@ -235,7 +235,7 @@ namespace UPC.Proyecto.SISPPAFUT
                                         objPartidoPronosticadoBE.C_Visita_Pts = objPartidoPronosticadoBE.C_Visita_Pts + 1;
                             }
                         }
-                        objPartidoPronosticadoBE.C_Visita_PosLiga = objTablaBC.ConsultarPosicionEquipoTabla(codLiga, codEquipoV);
+                        objPartidoPronosticadoBE.C_Visita_PosLiga = (int)objTablaBC.ConsultarPosicionEquipoTabla(codLiga, codEquipoV);
                         objPartidoPronosticadoBE.C_Visita_PosRankMund = objRankingBC.obtener_PosRanking(cDto.Fecha.Year, cDto.Fecha.Month, paisV, cDto.Equipo_visitante);
                         objPartidoPronosticadoBE.C_Visita_PromEdad = objEquipoBC.obtener_PromedioEquipoTitular(codEquipoV, codLiga);
                         objPartidoPronosticadoBE.C_Visita_QExpulsados = objEquipoBC.obtener_CantidadExpulsadosUltimoPartido(codUltimoPartidoVisita, codEquipoV, codLiga);
